action

+++ scripts/install.sh
@@ -280,11 +280,16 @@ if [ $OSNAME != "macos" ];then
 	mkdir -p /www/backup/site
 
 	if [ ! -d /www/server/mdserver-web ];then
-		curl --insecure -sSLo /tmp/master.tar.gz ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/master.tar.gz
-		cd /tmp && tar -zxvf /tmp/master.tar.gz
-		mv -f /tmp/bt_simple-master /www/server/mdserver-web
-		rm -rf /tmp/master.tar.gz
-		rm -rf /tmp/bt_simple-master
+		if [ -f "$PWD/cli.sh" ] && [ -d "$PWD/scripts" ]; then
+			echo "Local repository detected, copying files..."
+			cp -r $PWD /www/server/mdserver-web
+		else
+			curl --insecure -sSLo /tmp/master.tar.gz ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/master.tar.gz
+			cd /tmp && tar -zxvf /tmp/master.tar.gz
+			mv -f /tmp/bt_simple-master /www/server/mdserver-web
+			rm -rf /tmp/master.tar.gz
+			rm -rf /tmp/bt_simple-master
+		fi
 	fi
 
 	# install acme.sh
