+++ scripts/install_dev.sh
@@ -287,12 +287,17 @@ if [ $OSNAME != "macos" ];then
 	mkdir -p /www/backup/site
 
 	if [ ! -d /www/server/mdserver-web ];then
-		echo "downloading ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/dev.tar.gz"
-		curl --insecure -sSLo /tmp/dev.tar.gz ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/dev.tar.gz
-		cd /tmp && tar -zxvf /tmp/dev.tar.gz
-		mv -f /tmp/bt_simple-dev /www/server/mdserver-web
-		rm -rf /tmp/dev.tar.gz
-		rm -rf /tmp/bt_simple-dev
+		if [ -f "$PWD/cli.sh" ] && [ -d "$PWD/scripts" ]; then
+			echo "Local repository detected, copying files..."
+			cp -r $PWD /www/server/mdserver-web
+		else
+			echo "downloading ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/dev.tar.gz"
+			curl --insecure -sSLo /tmp/dev.tar.gz ${HTTP_PREFIX}github.com/clhome/bt_simple/archive/refs/heads/dev.tar.gz
+			cd /tmp && tar -zxvf /tmp/dev.tar.gz
+			mv -f /tmp/bt_simple-dev /www/server/mdserver-web
+			rm -rf /tmp/dev.tar.gz
+			rm -rf /tmp/bt_simple-dev
+		fi
 	fi
 
 	# install acme.sh
